requirements.txt
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.0
requests
django-allauth
dj-rest-auth
djangorestframework-simplejwt
python-dotenv
cryptography
oauth2client
